+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## HEAD
+
+- Add support for rails' `relative_url_root` config
+
 ## v1.9.2
 
 - Update rails dependency to allow rails 5